backend/requirements.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
psycopg2-binary